Moved/Renamed `I…TopicViewModel` classes
Moved `I…TopicViewModelCore` classes to `ViewModels\I…TopicViewModel`. The core made the interfaces unnecessarily wordy. And there are enough of them that it's more manageable to have them segmented into their own namespace now.
