+++ Ignia.Topics.Web.Mvc/TopicViewResult.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
+using Ignia.Topics.ViewModels;
 
 namespace Ignia.Topics.Web.Mvc {
 
@@ -36,7 +37,7 @@ namespace Ignia.Topics.Web.Mvc {
     ///   cref="ITopicViewModelCore.View"/> is unavailable, it is assumed to be the same as the <see
     ///   cref="ITopicViewModelCore.ContentType"/>.
     /// </remarks>
-    public TopicViewResult(ITopicViewModelCore viewModel) : base() {
+    public TopicViewResult(ITopicViewModel viewModel) : base() {
       ViewData.Model = viewModel;
       _contentType = viewModel.ContentType ?? "Page";
       _topicView = viewModel.View ?? _contentType;
