+++ Ignia.Topics/ViewModels/INavigationTopicViewModel{T}.cs
@@ -5,10 +5,10 @@
 \=============================================================================================================================*/
 using System.Collections.ObjectModel;
 
-namespace Ignia.Topics {
+namespace Ignia.Topics.ViewModels {
 
   /*============================================================================================================================
-  | INTERFACE: PAGE TOPIC VIEW MODEL CORE
+  | INTERFACE: PAGE TOPIC VIEW MODEL
   \---------------------------------------------------------------------------------------------------------------------------*/
   /// <summary>
   ///   Provides a generic data transfer topic for feeding views information about a navigation entry.
@@ -17,13 +17,13 @@ namespace Ignia.Topics {
   ///   No topics are expected to have a <c>Navigation</c> content type. Instead, implementers of this view model are expected
   ///   to manually construct instances.
   /// </remarks>
-  public interface INavigationTopicViewModelCore<T> : IPageTopicViewModelCore where T: INavigationTopicViewModelCore<T> {
+  public interface INavigationTopicViewModel<T> : IPageTopicViewModel where T: INavigationTopicViewModel<T> {
 
     /*==========================================================================================================================
     | PROPERTY: CHILDREN
     \-------------------------------------------------------------------------------------------------------------------------*/
     /// <summary>
-    ///   Represents a collection of child <see cref="INavigationTopicViewModelCore{T}"/> instances.
+    ///   Represents a collection of child <see cref="INavigationTopicViewModel{T}"/> instances.
     /// </summary>
     Collection<T> Children { get; set; }
 
@@ -34,7 +34,7 @@ namespace Ignia.Topics {
     ///   Determines if the current item is selected based on the provided <paramref name="uniqueKey"/>.
     /// </summary>
     /// <param name="uniqueKey">
-    ///   The unique key to compare against the current <see cref="INavigationTopicViewModelCore{T}"/>
+    ///   The unique key to compare against the current <see cref="INavigationTopicViewModel{T}"/>
     /// </param>
     bool IsSelected(string uniqueKey);
