+++ Ignia.Topics/ViewModels/IPageTopicViewModel.cs
@@ -5,23 +5,23 @@
 \=============================================================================================================================*/
 using Ignia.Topics.Mapping;
 
-namespace Ignia.Topics {
+namespace Ignia.Topics.ViewModels {
 
   /*============================================================================================================================
-  | INTERFACE: PAGE TOPIC VIEW MODEL CORE
+  | INTERFACE: PAGE TOPIC VIEW MODEL
   \---------------------------------------------------------------------------------------------------------------------------*/
   /// <summary>
   ///   Provides a generic data transfer topic for feeding views based on the page content type.
   /// </summary>
   /// <remarks>
   ///   <para>
-  ///     It is not strictly required that topic view models implement the <see cref="IPageTopicViewModelCore"/> interface for
-  ///     the default <see cref="TopicMappingService"/> to correctly identify and map <see cref="Topic"/>s to topic view models.
+  ///     It is not strictly required that topic view models implement the <see cref="IPageTopicViewModel"/> interface for the
+  ///     default <see cref="TopicMappingService"/> to correctly identify and map <see cref="Topic"/>s to topic view models.
   ///     That said, the interface does define properties that presentation infrastructure may expect. As a result, if it isn't
   ///     provided via the public interface then it will instead need to be defined in some other way.
   ///   </para>
   /// </remarks>
-  public interface IPageTopicViewModelCore : ITopicViewModelCore {
+  public interface IPageTopicViewModel : ITopicViewModel {
 
     /*==========================================================================================================================
     | PROPERTY: WEB PATH
@@ -38,7 +38,7 @@ namespace Ignia.Topics {
     ///   Gets or sets the Title attribute, which represents the friendly name of the topic.
     /// </summary>
     /// <remarks>
-    ///   While the <see cref="ITopicViewModelCore.Key"/> may not contain, for instance, spaces or symbols, there are no
+    ///   While the <see cref="ITopicViewModel.Key"/> may not contain, for instance, spaces or symbols, there are no
     ///   restrictions on what characters can be used in the title. For this reason, it provides the default public value for
     ///   referencing topics.
     /// </remarks>
