+++ Ignia.Topics/ViewModels/ITopicViewModel.cs
@@ -5,17 +5,17 @@
 \=============================================================================================================================*/
 using Ignia.Topics.Mapping;
 
-namespace Ignia.Topics {
+namespace Ignia.Topics.ViewModels {
 
   /*============================================================================================================================
-  | INTERFACE: TOPIC VIEW MODEL CORE
+  | INTERFACE: TOPIC VIEW MODEL
   \---------------------------------------------------------------------------------------------------------------------------*/
   /// <summary>
   ///   Provides a generic data transfer topic for feeding views.
   /// </summary>
   /// <remarks>
   ///   <para>
-  ///     It is not strictly required that topic view models implement the <see cref="ITopicViewModelCore"/> interface for the
+  ///     It is not strictly required that topic view models implement the <see cref="ITopicViewModel"/> interface for the
   ///     default <see cref="TopicMappingService"/> to correctly identify and map <see cref="Topic"/>s to topic view models.
   ///     That said, the interface does define properties that presentation infrastructure may expect. As a result, if it isn't
   ///     provided via the public interface then it will instead need to be defined in some other way.
@@ -23,11 +23,11 @@ namespace Ignia.Topics {
   ///   <para>
   ///     For instance, in the default MVC library, the <c>TopicViewResult</c> class requires that the <see
   ///     cref="Topic.ContentType"/> and <see cref="Topic.View"/> be supplied separately if they're not provided as part of a
-  ///     <see cref="ITopicViewModelCore"/>. The exact details of this will obviously vary based on the implementation of the
+  ///     <see cref="ITopicViewModel"/>. The exact details of this will obviously vary based on the implementation of the
   ///     presentation layer and any supporting libraries.
   ///   </para>
   /// </remarks>
-  public interface ITopicViewModelCore {
+  public interface ITopicViewModel {
 
     /*==========================================================================================================================
     | PROPERTY: KEY
